# web/openapi.v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ETÜTKOÇ API v2",
    "version": "0.1.0",
    "description": "Next.js BFF için JSON-only endpoint katmanı. Mobile için /api/v1 ayrıdır."
  },
  "paths": {
    "/api/v2/auth/login": {
      "post": {
        "tags": [
          "api-v2",
          "v2-auth"
        ],
        "summary": "V2 Login",
        "description": "E-posta + şifre → Set-Cookie ile JWT'ler kuruldu. Body'de sade user özeti.\n\napi_v1/auth.py:api_login ile aynı lockout + audit + generic error mesajı\npolitikası — kullanıcı sızıntısı (e-posta enumeration) önlenir.",
        "operationId": "v2_login_api_v2_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/app__routes__api_v2__auth__LoginOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginIn": {
        "properties": {
          "email": {
            "type": "string",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginIn"
      },
...
      "UserPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "role": {
            "type": "string",
            "enum": [
              "super_admin",
              "institution_admin",
              "teacher",
              "student",
              "parent"
            ],
            "title": "Role"
          },
          "institution_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Institution Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "must_change_password": {
            "type": "boolean",
            "title": "Must Change Password"
          },
          "last_login_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login At"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "full_name",
          "role",
          "is_active",
          "must_change_password"
        ],
        "title": "UserPublic",
        "description": "Kullanıcı kamuya açık alanları — password REDACTED."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "app__routes__api_v2__auth__LoginOut": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserPublic"
          },
          "must_change_password": {
            "type": "boolean",
            "title": "Must Change Password"
          }
        },
        "type": "object",
        "required": [
          "user",
          "must_change_password"
        ],
        "title": "LoginOut",
        "description": "Login response — token gövdede yok, cookie'de.\n\n`must_change_password=True` ise frontend kullanıcıyı /password/change'e\nyönlendirmeli; `get_current_user_v2` aksi halde 403 atar (R-001 azaltması)."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}